feat: implement platform-specific base directory retrieval for IVPN client
