+++ daemon/protocol/ivpnclient/file_locations_darwin.go
@@ -1,0 +1,9 @@
+//go:build darwin
+// +build darwin
+
+package ivpnclient
+
+func getPortFileBaseDir() (baseDir string, err error) {
+	baseDir = "/Library/Application Support/IVPN"
+	return baseDir, nil
+}
